% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bil_groups.R
\name{get_bil_groups}
\alias{get_bil_groups}
\title{get bilateral groups}
\usage{
get_bil_groups(term = 10)
}
\arguments{
\item{term}{the parliament term}
}
\value{
A data frame with all bilateral groups for a given term
}
\description{
get bilateral groups
}
