% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bil_groups_members.R
\name{get_bil_groups_members}
\alias{get_bil_groups_members}
\title{get members of bilateral groups}
\usage{
get_bil_groups_members(term = 10, group_id = NULL)
}
\arguments{
\item{term}{the parliament term}

\item{group_id}{the id (or ids as a vector) of the bilateral group. If left to NULL will get data for all bilateral groups}
}
\value{
A data frame with members of bilateral groups
}
\description{
get members of bilateral groups
}
