% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_committees.R
\name{get_comittees}
\alias{get_comittees}
\title{get all committees for a given term}
\usage{
get_comittees(term = 10)
}
\arguments{
\item{term}{the parliament term}
}
\value{
A data frame with all committees in a given parliament term
}
\description{
get all committees for a given term
}
