% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_interpellations.R
\name{get_interpellations}
\alias{get_interpellations}
\title{get all MP interpellations for a given term}
\usage{
get_interpellations(term = 10)
}
\arguments{
\item{term}{the parliament term}
}
\value{
A data frame with all interpellations
}
\description{
get all MP interpellations for a given term
}
