% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_legislation.R
\name{get_legislation}
\alias{get_legislation}
\title{get basic information on legislation in a given term}
\usage{
get_legislation(term = 10, n = 50)
}
\arguments{
\item{term}{the parliament term}

\item{n}{the number of legal acts to return (defaults to 50)}
}
\value{
A data frame with information on legal acts
}
\description{
get basic information on legislation in a given term
}
