% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mp_votes.R
\name{get_mp_votes}
\alias{get_mp_votes}
\title{get information how MPs voted in particular votes}
\usage{
get_mp_votes(term = 10, proceeding = NULL, vote_id = NULL)
}
\arguments{
\item{term}{the parliament term}

\item{proceeding}{specific proceeding to pull votes from. Will get all proceedings if left to NULL}

\item{vote_id}{specific id of a vote to pull how each MP voted. Will get all votes if left to NULL}
}
\value{
A data frame with information how each MP voted in given votes
}
\description{
get information how MPs voted in particular votes
}
