% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mps.R
\name{get_mps}
\alias{get_mps}
\title{get basic information on MPs in a given term}
\usage{
get_mps(term = 10)
}
\arguments{
\item{term}{the parliament term}
}
\value{
A data frame with all MPs for a given term
}
\description{
get basic information on MPs in a given term
}
