% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parties.R
\name{get_parties}
\alias{get_parties}
\title{get basic information on parties}
\usage{
get_parties(term = 10)
}
\arguments{
\item{term}{the parliament term}
}
\value{
A data frame with information on parties
}
\description{
get basic information on parties
}
