% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proceedings.R
\name{get_proceedings}
\alias{get_proceedings}
\title{get proceedings pf polish sejm}
\usage{
get_proceedings(term = 10)
}
\arguments{
\item{term}{the parliament term}
}
\value{
A data frame with all proceedings for a given term
}
\description{
get proceedings pf polish sejm
}
