% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_questions.R
\name{get_questions}
\alias{get_questions}
\title{get information on written questions by MPs}
\usage{
get_questions(term = 10)
}
\arguments{
\item{term}{the parliament term}
}
\value{
A data frame with all questions
}
\description{
get information on written questions by MPs
}
