% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_votes.R
\name{get_votes}
\alias{get_votes}
\title{get information on votes per proceeding}
\usage{
get_votes(term = 10, proceeding = NULL)
}
\arguments{
\item{term}{the parliament term}

\item{proceeding}{specific proceeding to pull votes for. Will get all proceedings if left to NULL}
}
\value{
A data frame with all interpellations
}
\description{
get information on votes per proceeding
}
